feat: OrderRemoteSource를 Retrofit, Moshi로 구현

+++ app/src/main/java/woowacourse/shopping/ui/order/uistate/DiscountPolicyUIState.kt
@@ -4,7 +4,7 @@ import woowacourse.shopping.domain.order.DiscountPolicy
 
 data class DiscountPolicyUIState(
     val name: String,
-    val discountRate: Int,
+    val discountRate: Double,
     val discountPrice: Int
 ) {
