+++ domain/src/main/java/woowacourse/shopping/domain/order/DiscountPolicy.kt
@@ -2,6 +2,6 @@ package woowacourse.shopping.domain.order
 
 data class DiscountPolicy(
     val name: String,
-    val discountRate: Int,
+    val discountRate: Double,
     val discountPrice: Int
 )
